+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 - Discounted land will again be awarded when buildings are lost to invasion
 - Infamy now increases mana production at the same rate as gems/lumber/ore
 - Infamy gain now based on chance of success instead of relative ratio
+- Hourly infamy decay reduced by 25% while in the Shadow League
 - Spies and wizards lost due to failed black ops when targeting players with much higher ratios have been capped around 2 SPA/WPA (info ops already capped when over 0.6 SPA)
 - Magic snare now does a minimum of 1% damage after resilience (from 1.5% before resilience)
 - Wizard strength refresh rate increased by 1% if you would remain under 30% the following tick